requirements.txt
tensorflow
streamlit
numpy
pillow
opencv-python
scikit-image
matplotlib
lime
fpdf
